Moved PostsPageTopBarProtocol to PostsPageTopBarLiquidGlass.swift
